SignUpコンポーネントの分割が誤っていたので修正
