+++ frontend/src/features/auth/components/SignUpContainer.tsx
@@ -1,0 +1,21 @@
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+
+import { SignUpForm } from "@/features/auth/components/SignUpForm";
+import { signUpSchema, type SignUpSchema } from "@/features/auth/types";
+import { createClient } from "@/lib/supabase/client";
+
+const SignUpContainer = () => {
+  const supabase = createClient();
+  const methods = useForm<SignUpSchema>({
+    resolver: zodResolver(signUpSchema),
+  });
+
+  const onSubmit = async (data: SignUpSchema) => {
+    await supabase.auth.signUp(data);
+  };
+
+  return <SignUpForm methods={methods} onSubmit={onSubmit} />;
+};
+
+export default SignUpContainer;
